requirements.txt
facebook-business>=14.0.0
python-dotenv>=0.19.0
requests>=2.28.0
pandas>=1.3.0
tabulate>=0.8.9